+++ R/8_summaries.R
@@ -3,7 +3,7 @@
 #' These functions are used to summarise the simulation outputs, typically by joining the optimal scores with a matching schedule.
 #'
 #' @param optimal_scores a dataframe of optimized lineups as created by `ffs_optimize_lineups()`
-#' @param schedules a dataframe of schedules as created by `ffs_build_schedules()`
+#' @param schedules a dataframe of schedules as created by `ffs_build_schedules()` or `ffs_actual_schedules()`
 #'
 #' @seealso `vignette("Custom Simulations")` for example usage
 #'
@@ -18,7 +18,6 @@
 #' summary_season
 #' summary_simulation <- ffs_summarise_simulation(summary_season)
 #' summary_simulation
-#'
 #' }
 #'
 #' @rdname ffs_summaries
@@ -33,7 +32,7 @@ ffs_summarise_week <- function(optimal_scores, schedules) {
   checkmate::assert_data_frame(schedules)
   assert_columns(
     schedules,
-    c("season", "week", "team", "opponent")
+    c("league_id","franchise_id","opponent_id","season", "week")
   )
 
   scores <- optimal_scores %>%
@@ -49,16 +48,18 @@ ffs_summarise_week <- function(optimal_scores, schedules) {
   matchups <- schedules %>%
     dplyr::left_join(scores %>%
       dplyr::rename("team_score" = "actual_score"),
-    by = c("team" = "schedule_id", "season", "week")
+      by = c("league_id","franchise_id", "season", "week")
     ) %>%
     dplyr::left_join(scores %>%
       dplyr::select(
         "opponent_score" = "actual_score",
-        "schedule_id",
+        "opponent_id" = "franchise_id",
         "opponent_name" = "franchise_name",
-        "season", "week"
+        "league_id",
+        "season",
+        "week"
       ),
-    by = c("opponent" = "schedule_id", "season", "week")
+      by = c("league_id","opponent_id", "season", "week")
     ) %>%
     dplyr::mutate(
       result = dplyr::case_when(
@@ -84,7 +85,8 @@ ffs_summarise_week <- function(optimal_scores, schedules) {
       "allplay_pct",
       "league_id",
       "franchise_id",
-      "optimal_lineup"
+      "optimal_player_id",
+      "optimal_player_score"
     )
 
   return(matchups)
@@ -100,7 +102,8 @@ ffs_summarise_season <- function(summary_week) {
   checkmate::assert_data_frame(summary_week)
   assert_columns(
     summary_week,
-    c("season", "league_id", "franchise_id", "franchise_name",
+    c(
+      "season", "league_id", "franchise_id", "franchise_name",
       "result", "allplay_wins", "allplay_games",
       "team_score", "opponent_score", "optimal_score"
     )
@@ -133,7 +136,8 @@ ffs_summarise_simulation <- function(summary_season) {
   checkmate::assert_data_frame(summary_season)
   assert_columns(
     summary_season,
-    c("league_id", "franchise_id", "franchise_name",
+    c(
+      "league_id", "franchise_id", "franchise_name",
       "h2h_wins", "h2h_winpct", "allplay_wins", "allplay_winpct",
       "points_for", "points_against", "potential_points"
     )